+++ Mail/DMOrganizePreferencesViewController.m
@@ -140,7 +140,7 @@ static const NSInteger kAccountsTableViewTag = 101;
 
 - (void)tableViewSelectionDidChange:(NSNotification *)notification {
 	NSTableView *tableView = (NSTableView *)notification.object;
-	if (tableView.tag == kAccountsTableViewTag && tableView.selectedRow < PSTAccountManager.defaultManager.accounts.count) {
+	if (tableView.tag == kAccountsTableViewTag && tableView.selectedRow < (NSInteger)PSTAccountManager.defaultManager.accounts.count) {
 		self.selectedAccount = PSTAccountManager.defaultManager.accounts[tableView.selectedRow];
 	} else {
 		NSRect selectedRect = CGRectOffset([tableView rectOfRow:tableView.selectedRow], 0, [tableView.enclosingScrollView documentVisibleRect].origin.y + 10);
